Refactor listener.KiwiTCMS to use tcms_api.plugin_helpers.Backend
this makes the behavior more in-sync with existing plugins and
obsoletes the zealand.library module which is very similar.

There is some nice extra integration from the original
implementation which we'd like to keep around! This is done by
overriding the Backend class and implementing these extra bits.

This also reveals the need for some changes in tcms-api marked
here as todo: comments (which will make pylint unhappy).

Refs #4 b/c I've updated README as well.