Refactor listener.KiwiTCMS to use tcms_api.plugin_helpers.Backend
this makes the behavior more in-sync with existing plugins and
obsoletes the zealand.library module which is very similar.

There is some nice extra integration from the original
implementation which we keep as much as possible! This is done by
overriding the Backend class and implementing these extra bits.

Refs #4 b/c I've updated README as well.

Also document how to override and extend with a simple example
inspired by the original implementation.